Conflicts:
	camera.cpp
	rayon.cpp

+++ camera.cpp
@@ -15,7 +15,6 @@ Camera::Camera(const Vector3f& pOr, const Vector3f& pAt, int l, int h, const std
     _w = pAt - pOr;
 
     _lw = _w.norm();
-
     _w.normalize();
 
     if(_w == Vector3f(0,0,1) || _w == Vector3f(0,0,-1))
@@ -40,11 +39,11 @@ Vector3f Camera::vecScreen(int i, int j) const
         exit(-1);
     }
 
-    float ti = i / (_lu*2-1);
-    float tj = j / (_lv*2-1);
+    float ti = ((float)i) / (_lu*2 - 1);
+    float tj = ((float)j) / (_lv*2 - 1);
 
 
-    return _w*_lw + ( (1 - ti)*(-_lu)+(ti*_lu) )*_u + ( (1-tj)*(-_lv)+(tj*_lv) )*_v;
+    return _w*_lw + ( (1.0 - ti)*(-_lu)+(ti*_lu) )*_u + ( (1.0-tj)*(-_lv)+(tj*_lv) )*_v;
 }
 
 Vector3f Camera::pointScreen(int i, int j) const
